The health bar now reflects the current health of the hero. The hero will lose one life per second if he is on fire. There is no game over screen yet though.

+++ CS185TD/src/com/touchspin/td/g.java
@@ -29,7 +29,7 @@ public class g {
 	public float playerFriction; // The friction of the player for physics
 	public boolean fire;	// Hero is on fire
 	public String currentBallType;
-	public float playerHealth;  // Health of the player. 1=full health. 
+	public int playerHealth;  // Health of the player. 39=full health. 
 	public Hero hero;
 	public HUD hud;
